Rename usePlotBandLine to usePlotBandLineLifecycle
