+++ packages/react-jsx-highcharts/src/components/PlotBandLine/PlotLine.js
@@ -1,10 +1,10 @@
 import React, { memo } from 'react';
 import PropTypes from 'prop-types';
 import PlotBandLineContext from './PlotBandLineContext';
-import usePlotBandLine from './usePlotBandLine';
+import usePlotBandLineLifecycle from './usePlotBandLineLifecycle';
 
 const PlotLine = memo(props => {
-  const plotline = usePlotBandLine(props, 'plotLines');
+  const plotline = usePlotBandLineLifecycle(props, 'plotLines');
 
   const { children } = props;
